+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Fixed
 
 * Equations and adjustments for the new manager
+* Adjust TriggerType for OpenAI and Typebot integrations
 
 # 2.0.2-beta (2024-07-18 21:33)